Update Audio_Pipeline.v
Switch the input clock of the pcm_to_pwm module. By increasing the clock rate in this module the audio quality is much better. Also removed bit-masking from the pdm to pcm module in order to maintain as much of the original sound as possible. Small white noise is still audible, but the overall sound quality is much better than before.