Update Audio_Pipeline.v
Changed the mic_clk from 4.16 Mhz to 4.55 Mhz. This increases the power usage of the mic but also ensures that the MEMS microphone on the FPGA is operating in high-power mode. This change removed the static sound. I also changed the frequency of the PWM output to 96Khz from 48Khz. There is no noticeable difference in audio quality, however, this system is now closer to being "Hi-Fi".